// src/ts/commons/constants/enums.ts
export enum ProgrammingLanguage {
  CSS = 'CSS',
  HTML = 'HTML',
  JS = 'JavaScript',
  NodeJS = 'Node.js',
  OCaml = 'OCaml',
  Python = 'Python',
  ReasonML = 'ReasonML',
  Elixir = 'Elixir',
}

export enum Library {
  BsCSS = 'bs-css',
  BsFetch = 'bs-fetch',
  BsJSON = 'bs-json',
  BabelJS = 'Babel',
  CSSInJS = 'CSS in JS',
  DraftJS = 'DraftJS',
  Electron = 'Electron',
  FoundationForSites = 'Foundation for Sites',
  FlowTyped = 'flow-typed',
  GulpJS = 'Gulp',
  Imagemin = 'imagemin',
  Jest = 'Jest',
  JQuery = 'jQuery',
  MediaElementJS = 'MediaElement.js',
  MomentJS = 'moment.js',
  NPM = 'npm',
  ParcelJS = 'Parcel',
  Phantomas = 'phantomas',
  PraytimeJS = 'praytime.js',
  ReactJS = 'React.js',
  ReactNative = 'React Native',
  ReasonReact = 'ReasonReact',
  ReactChartJS2 = 'ReactChart.js 2',
  Reformality = 're-formality',
  Redux = 'Redux',
  ReduxSaga = 'Redux Saga',
  Rematch = 'Rematch',
  Sass = 'Sass',
  StyledComponents = 'styled-components',
  TypeScript = 'TypeScript',
  Webpack = 'Webpack',
  Yarn = 'Yarn',
}

export enum Tool {
  Git = 'Git',
  TMUX = 'TMUX',
  LaTeX = 'LaTeX',
  Vim = 'vim',
}

export enum ProjectType {
  Web = 'Web',
  Mobile = 'Mobile',
  Infrastructure = 'Infrastructure',
  Desktop = 'Desktop',
}